requirements.txt
streamlit
google-generativeai
pandas
Pillow
watchdog
